% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bw180905_53_10hz}
\alias{bw180905_53_10hz}
\title{Blue whale 10 Hz depth and acceleration}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 72001 rows and 3 columns.
}
\usage{
bw180905_53_10hz
}
\description{
Note to self: add PRH
}
\keyword{datasets}
