% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bw180905_53_400hz}
\alias{bw180905_53_400hz}
\title{Blue whale 400 Hz acceleration}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2880001 rows and 4 columns.
}
\usage{
bw180905_53_400hz
}
\description{
Blue whale 400 Hz acceleration
}
\keyword{datasets}
