% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bw180905_53_elg}
\alias{bw180905_53_elg}
\title{Blue whale BCG eligible regions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 4 columns.
}
\usage{
bw180905_53_elg
}
\description{
Blue whale BCG eligible regions
}
\keyword{datasets}
