% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_beats.R
\name{choose_peaks}
\alias{choose_peaks}
\title{Choose peaks to keep}
\usage{
choose_peaks(x)
}
\arguments{
\item{x}{\verb{[numeric]} peak prominences}
}
\value{
\verb{[logical]}
}
\description{
Uses a heuristic approach to keep the most prominent peaks
}
