% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corky_10hz}
\alias{corky_10hz}
\title{Killer whale 10 Hz body orientation and jerk}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6908 rows and 5 columns.
}
\usage{
corky_10hz
}
\description{
Note to self: add depth, Aw. Remove jerk.
}
\keyword{datasets}
