% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corky_400hz}
\alias{corky_400hz}
\title{Killer whale 400 Hz acceleration}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 617728 rows and 4 columns.
}
\usage{
corky_400hz
}
\description{
Killer whale 400 Hz acceleration
}
\keyword{datasets}
