% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corky_ecg}
\alias{corky_ecg}
\title{Killer whale ECG}
\format{
An object of class \code{data.frame} with 180057 rows and 3 columns.
}
\usage{
corky_ecg
}
\description{
Killer whale ECG
}
\keyword{datasets}
