% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecg_beats}
\alias{ecg_beats}
\title{Killer whale heart beats recorded by ECG}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1075 rows and 3 columns.
}
\usage{
ecg_beats
}
\description{
Killer whale heart beats recorded by ECG
}
\keyword{datasets}
