% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_beats.R
\name{filter_acc}
\alias{filter_acc}
\title{Apply bandpass filter to acceleration}
\usage{
filter_acc(acc, fs, upper)
}
\arguments{
\item{acc}{\verb{[numeric(n,1)]} or \verb{[numeric(n,3)]} Longitudinal acceleration.
Either surge only or tri-axial.}

\item{fs}{\verb{[numeric(1)]} sampling frequency in Hz}

\item{upper}{\verb{[numeric(1)]} upper cutoff for bandpass filter (in Hz)}
}
\value{
\verb{[numeric]} filtered acceleration
}
\description{
Apply bandpass filter to acceleration
}
