% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_beats.R
\name{find_beats}
\alias{find_beats}
\title{Find beats in BCG}
\usage{
find_beats(cue, fs_hz, minperiod)
}
\arguments{
\item{cue}{\verb{[numeric]} signal used for heart beat cue.}

\item{fs_hz}{\verb{[numeric(1)]} Sampling frequency of surge in Hz}

\item{minperiod}{\verb{[numeric(1)]} Minimum period between beats (in seconds)}
}
\value{
\verb{[logical]} same length as cue indicating heart beats
}
\description{
Find beats in BCG
}
