% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_beats.R
\name{find_beats_lsr}
\alias{find_beats_lsr}
\title{Find beats in BCG with local surge range cue}
\usage{
find_beats_lsr(surge, window_s = 0.25, fs_hz = 400)
}
\arguments{
\item{surge}{Filtered surge signal}

\item{window_s}{Width of search window in seconds (0.25 s by default)}

\item{fs_hz}{Sampling frequency of surge in Hz (400 Hz by default)}
}
\value{
A logical vector the same length as surge indicating a heart beat
}
\description{
Find beats in BCG with local surge range cue
}
