% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{jerk}
\alias{jerk}
\title{Norm of the jerk vector}
\usage{
jerk(A, fs, p = 3, n = 11)
}
\arguments{
\item{A}{Tri-axial acceleration (n x 3 matrix)}

\item{fs}{Sampling rate (Hz)}

\item{p}{Order of Savitzky-Golay filter (3 by default)}

\item{n}{Window size of Savitzky-Golay filter (11 by default)}
}
\value{
Norm of the jerk vector of A
}
\description{
Uses Savitzky-Golay filtering to de-noise before differentiating acceleration
}
