% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_beats.R
\name{local_range}
\alias{local_range}
\title{Local range of signal}
\usage{
local_range(x, k)
}
\arguments{
\item{x}{Original signal}

\item{k}{Window width}
}
\value{
Smoothed local range of x
}
\description{
Local range of signal
}
