% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{njerk}
\alias{njerk}
\title{Norm of the jerk vector}
\usage{
njerk(A, fs)
}
\arguments{
\item{A}{Tri-axial acceleration (n x 3 matrix)}

\item{fs}{Sampling rate (Hz)}
}
\value{
Norm of the jerk vector of A
}
\description{
Norm of the jerk vector
}
