% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dives.R
\name{normalize_dives}
\alias{normalize_dives}
\title{Normalize dive time}
\usage{
normalize_dives(dive_id)
}
\arguments{
\item{dive_id}{A factor of dive id's, as from \code{split_dives()}.}
}
\value{
A numeric vector, the same length as \code{dive_id}.
}
\description{
Normalize the time in each dive to \verb{[0, 1]} where 0 is start of dive and 1 is
the end.
}
