% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{peak_prominences}
\alias{peak_prominences}
\title{Find prominences of peaks}
\usage{
peak_prominences(x, peaks)
}
\arguments{
\item{x}{Vector of numeric values}

\item{peaks}{Indices of peak locations}
}
\value{
Prominences of peaks in \code{x} at indices \code{peaks}
}
\description{
Find prominences of peaks
}
