% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{pspectrum}
\alias{pspectrum}
\title{Estimate power spectral density}
\usage{
pspectrum(x, fs_hz)
}
\arguments{
\item{x}{\verb{[numeric]} signal}

\item{fs_hz}{\verb{[numeric(1)]} sampling rate (Hz)}
}
\value{
\verb{[spec]} see \code{psd::pspectrum}
}
\description{
Estimate power spectral density
}
