% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{s2n}
\alias{s2n}
\title{Calculate signal-to-noise ratio from power spectral density}
\usage{
s2n(psd, signal_bpm)
}
\arguments{
\item{psd}{\verb{[spec]} power spectral density}

\item{signal_bpm}{\verb{[numeric(2)]} signal band (bpm)}
}
\value{
\verb{[numeric(1)]}
}
\description{
Calculate signal-to-noise ratio from power spectral density
}
