% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Shannon entropy}
\usage{
shannon_entropy(x)
}
\arguments{
\item{x}{\verb{[numeric] vector or matrix}}
}
\value{
Shannon entropy of x
}
\description{
Defined as -sum(|x| * log(|x|))
}
