% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dives.R
\name{split_dives}
\alias{split_dives}
\title{Split dive profile into dives}
\usage{
split_dives(dt, depth, surface = 0, min_depth = 0, min_dur = 0)
}
\arguments{
\item{dt}{Datetime (POSIXct)}

\item{depth}{Depth (numeric)}

\item{surface}{Threshold for surface activity (numeric(1))}

\item{min_depth}{Minimum depth for a dive (numeric(1))}

\item{min_dur}{Minimum duration of a dive (numeric(1), in seconds)}
}
\value{
A factor, the same length as \code{depth}, with levels corresponding
to individual dives. Inter-dive periods marked as NA.
}
\description{
Split dive profile into dives
}
